refactor: extract converter postprocess + split tests into separate files
- Extract postprocess() and group_segments() from mod.rs to postprocess.rs (101L)
- Split ~830 lines of mod.rs tests into tests/{basic,nbest,history,grouping,bench}.rs
- Move ~570 lines of reranker.rs tests to tests/reranker.rs

Result: mod.rs 1,064→138 lines, reranker.rs 700→129 lines.
All test paths now clearly organized by concern.
